requirements.txt
langchain
openai
unstructured
pinecone-client
pypdf
pandas 
numpy
python-dotenv
tiktoken
streamlit